refactor(suggestions): extrai helpers compartilhados p/ reuso
Move trechos genéricos do fluxo de sugestões para libs reutilizáveis, sem
mudança de comportamento (mesma sanitização, Turnstile e montagem de Discord):

- lib/text-schemas.ts: sanitizeText/shortText (re-exportados em suggestion-schemas)
- lib/discord.ts: DiscordWebhookPayload/DiscordSender/truncate
- lib/hygraph-response.ts: readIdFromResponse + tipo HygraphRequest
- lib/form-errors.ts: extractIssues
- server/utils/turnstile.ts: verifyTurnstile (auto-importado pelo Nitro)

Prepara o terreno para o canal de feedback reaproveitar o mesmo pipeline.
